*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3640 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func_util.c
@@ -819,7 +819,7 @@ Var *f_cscale( Var *v )
 Var *f_clabel( Var *v )
 {
 	char *l[ 3 ] = { NULL, NULL, NULL };
-	long lengths[ 3 ] = { 1, 1, 1 };
+	size_t lengths[ 3 ] = { 1, 1, 1 };
 	int shm_id;
 	long len = 0;                    /* total length of message to send */
 	void *buf;
@@ -1080,7 +1080,7 @@ Var *f_display( Var *v )
 {
 	DPoint *dp;
 	int shm_id;
-	size_t len = 0;                    /* total length of message to send */
+	long len = 0;                     /* total length of message to send */
 	void *buf;
 	char *ptr;
 	int nsets;
